+++ CHANGELOG.md
@@ -1,4 +1,14 @@
-# [0.4.4] - New Extensions
+# [0.5.0] - Breaking Changes & New extensions
+
+- All methods will now trim the `String` before operations.
+- `toSnakeCase` : Converts a `String` to snake case format.
+
+## [0.4.5] - New extensions
+
+- `addAfter` : Adds a `String` after the first occurence of a specific pattern.
+- `addBefore` : Adds a `String` before the first occurence of a specific pattern.
+
+## [0.4.4] - New Extensions
 
 - `removeAfter` : Removes everything after the first occurence of a specific pattern.